Move methods to new class
